Move reporter middleware into package reporter.
