+++ empire/pkg/httpx/middleware/reporter_test.go
@@ -1,36 +1,41 @@
-package reporter
+package middleware
 
 import (
+	"errors"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
 	"github.com/remind101/empire/empire/pkg/httpx"
+	"github.com/remind101/empire/empire/pkg/reporter"
 	"golang.org/x/net/context"
 )
 
-func TestMiddleware(t *testing.T) {
-	var called bool
+func TestReporter(t *testing.T) {
+	var (
+		called  bool
+		errBoom = errors.New("boom")
+	)
 
-	r := ReporterFunc(func(ctx context.Context, err error) error {
+	r := reporter.ReporterFunc(func(ctx context.Context, err error) error {
 		called = true
 
-		if err != ErrFake {
-			t.Fatalf("err => %v; want %v", err, ErrFake)
+		if err != errBoom {
+			t.Fatalf("err => %v; want %v", err, errBoom)
 		}
 
 		return nil
 	})
 
 	h := httpx.HandlerFunc(func(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
-		_, ok := FromContext(ctx)
+		_, ok := reporter.FromContext(ctx)
 		if !ok {
 			t.Fatal("Expected the handler to be added to the context")
 		}
 
-		return ErrFake
+		return errBoom
 	})
-	m := &Middleware{
+	m := &Reporter{
 		Reporter: r,
 		handler:  h,
 	}
@@ -41,8 +46,8 @@ func TestMiddleware(t *testing.T) {
 
 	err := m.ServeHTTPContext(ctx, resp, req)
 
-	if err != ErrFake {
-		t.Fatalf("err => %v; want %v", err, ErrFake)
+	if err != errBoom {
+		t.Fatalf("err => %v; want %v", err, errBoom)
 	}
 
 	if !called {
